* ADS/LDAP: periodic import of account into SQL database - import from users, groups and memberships - using (simple) paged result to kope with result size limitation from LDAP servers - incremental sync uses modification time - async import job and logging --> ToDo: deleting of accounts and testing with LDAP

+++ api/src/Accounts/Sql.php
@@ -188,9 +188,10 @@ class Sql
 	 * If no account_id is set in data the account is added and the new id is set in $data.
 	 *
 	 * @param array $data array with account-data
-	 * @return int/boolean the account_id or false on error
+	 * @param bool $force_create true: do NOT check with frontend, if account exists
+	 * @return int|false the account_id or false on error
 	 */
-	function save(&$data)
+	function save(&$data, $force_create=false)
 	{
 		$to_write = $data;
 		unset($to_write['account_passwd']);
@@ -207,7 +208,7 @@ class Sql
 			$to_write['account_lastpwd_change'] = time();
 		}
 		if ($data['mustchangepassword'] == 1) $to_write['account_lastpwd_change']=0;
-		if (!(int)$data['account_id'] || !$this->id2name($data['account_id']))
+		if ($force_create || !(int)$data['account_id'] || !$this->id2name($data['account_id']))
 		{
 			if ($to_write['account_id'] < 0) $to_write['account_id'] *= -1;
 
@@ -268,7 +269,10 @@ class Sql
 	}
 
 	/**
-	 * Delete one account, deletes also all acl-entries for that account
+	 * Delete one account
+	 *
+	 * Does NOT delete acl-entries and memberships, use Acl::delete_account($account_id) for that!
+	 * Users need to be deleted via admin_cmd_delete_account, to ensure proper data removal.
 	 *
 	 * @param int $account_id numeric account_id
 	 * @return boolean true on success, false otherwise
@@ -277,13 +281,11 @@ class Sql
 	{
 		if (!(int)$account_id) return false;
 
-		$contact_id = $this->id2name($account_id,'person_id');
-
 		if (!$this->db->delete($this->table,array('account_id' => abs($account_id)),__LINE__,__FILE__))
 		{
 			return false;
 		}
-		if ($contact_id)
+		if ($account_id > 0 && ($contact_id = $this->id2name($account_id,'person_id')))
 		{
 			if (!isset($this->contacts)) $this->contacts = new Api\Contacts();
 			$this->contacts->delete($contact_id,false);	// false = allow to delete accounts (!)
